+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.1.63] - 2020-08-25
+### Added
+- The `corner back button label` screen part.
+### Changed
+- Upgraded bootstrap-fileupload.
+### Fixed
+- The `progress` modifier did not behave as described in the documentation.
+
 ## [1.1.62] - 2020-08-23
 ### Added
 - The modifiers `include attachment notice`, `include download tab`,